+++ selftests/config/WebTest.dtd
@@ -1,5 +1,5 @@
-<?xml version="1.0" encoding="us-ascii"?><!--Generated on 04-Aug-2017 21:13 for build development-20170804.2112-->
-<!ENTITY % step "(pdfToLinksFilter | selectWebClient | pdfVerifyPageCount | excelStoreCellValue | verifyCheckbox | ifStep | setRadioButton | forceInputFieldAttribute | replaceFilter | verifyNoDialogs | pdfVerifyBookmark | storeJsonPath | storeRegEx | verifyContent | not | confirm | verifyLinks | applyFilters | pdfVerifyInfoProperty | pdfVerifyField | verifyText | pdfVerifyFont | emailStoreHeader | verifyElement | pdfVerifyCheckboxField | emailStoreMessageId | normalizeXml | excelFindRow | groovy | contentStripper | verifyElementText | setInputField | reportLinks | emailSetConfig | pdfDecryptDocument | storeURLEncoded | verifyHeader | verifyRadioButton | invoke | retry | pdfVerifyEncryptionPermissions | group | pdfVerifyEncryptionStrength | clickLink | selectWindow | pdfToTextFilter | selectForm | pdfVerifyTitle | closeWindow | testInfo | expectDialog | dataDriven | alert | storeInputFieldAttribute | reportSite | pdfToBookmarksFilter | pdfVerifyEncryption | storeCookie | storeDigest | previousResponse | emailMessageStructureFilter | storeProperty | timedGroup | verifyDocumentURL | enableJavaScript | mouseOut | steps | matchLinesFilter | setFileField | storeURLDecoded | pdfVerifyLink | verifyProperty | verifyImages | verifyTextarea | repeat | verifyJsonPath | verifyLinkedContent | config | prompt | storeHeader | webtest | tagStripper | dragAndDrop | pdfVerifyFieldCount | verifyCookie | lineSeparatorFilter | excelStructureFilter | followFrame | pdfToFontsFilter | emailStorePartCount | verifyInputField | countWebtestResults | storeRandom | pdfVerifyTextField | selectLinesFilter | expectDialogs | storeLength | emailDeleteMessage | excelSelectSheet | antWrapper | storeLinkParameter | excelVerifyCellStyle | verifyTitle | setCheckbox | storeElementAttribute | excelVerifyCellSum | verifyXPath | pdfVerifyText | clickElement | verifySelectField | forceHiddenInputField | emailStorePartFilename | clickButton | emailStoreMessageType | setSelectField | scriptStep | storeXPath | emailMessageContentFilter | storeResponseCode | excelVerifyCellValue | pdfVerifyReadOnlyField | emailStoreMessageCount | mouseOver | pdfVerifyEncryptionProperty)" >
+<?xml version="1.0" encoding="us-ascii"?><!--Generated on 22-Dec-2017 16:10 for build development-20171222.1609-->
+<!ENTITY % step "(pdfToLinksFilter | selectWebClient | pdfVerifyPageCount | excelStoreCellValue | verifyCheckbox | ifStep | setRadioButton | forceInputFieldAttribute | replaceFilter | verifyNoDialogs | pdfVerifyBookmark | storeJsonPath | storeRegEx | verifyContent | not | confirm | verifyLinks | applyFilters | pdfVerifyInfoProperty | pdfVerifyField | verifyText | pdfVerifyFont | emailStoreHeader | verifyElement | pdfVerifyCheckboxField | emailStoreMessageId | normalizeXml | excelFindRow | groovy | contentStripper | verifyElementText | setInputField | reportLinks | emailSetConfig | pdfDecryptDocument | storeURLEncoded | verifyHeader | verifyRadioButton | invoke | retry | pdfVerifyEncryptionPermissions | group | pdfVerifyEncryptionStrength | clickLink | selectWindow | pdfToTextFilter | selectForm | pdfVerifyTitle | closeWindow | testInfo | expectDialog | dataDriven | alert | storeInputFieldAttribute | reportSite | invokeWithHeaders | pdfToBookmarksFilter | pdfVerifyEncryption | storeCookie | storeDigest | previousResponse | emailMessageStructureFilter | storeProperty | timedGroup | verifyDocumentURL | enableJavaScript | mouseOut | steps | matchLinesFilter | setFileField | storeURLDecoded | pdfVerifyLink | verifyProperty | verifyImages | verifyTextarea | repeat | verifyJsonPath | verifyLinkedContent | config | prompt | storeHeader | webtest | tagStripper | dragAndDrop | pdfVerifyFieldCount | verifyCookie | lineSeparatorFilter | excelStructureFilter | followFrame | pdfToFontsFilter | emailStorePartCount | verifyInputField | countWebtestResults | storeRandom | pdfVerifyTextField | selectLinesFilter | expectDialogs | storeLength | emailDeleteMessage | excelSelectSheet | antWrapper | storeLinkParameter | excelVerifyCellStyle | verifyTitle | setCheckbox | storeElementAttribute | excelVerifyCellSum | verifyXPath | pdfVerifyText | clickElement | verifySelectField | forceHiddenInputField | emailStorePartFilename | clickButton | emailStoreMessageType | setSelectField | scriptStep | storeXPath | emailMessageContentFilter | storeResponseCode | excelVerifyCellValue | pdfVerifyReadOnlyField | emailStoreMessageCount | mouseOver | pdfVerifyEncryptionProperty)" >
 <!ELEMENT step EMPTY >
 <!ELEMENT task EMPTY >
 <!ELEMENT condition EMPTY >
@@ -442,6 +442,21 @@
 	depth CDATA #IMPLIED
 	description CDATA #IMPLIED
 	file CDATA #IMPLIED
+>
+<!ELEMENT invokeWithHeaders EMPTY >
+<!ATTLIST invokeWithHeaders
+	content CDATA #IMPLIED
+	contentFile CDATA #IMPLIED
+	contentType CDATA #IMPLIED
+	description CDATA #IMPLIED
+	method CDATA #IMPLIED
+	password CDATA #IMPLIED
+	save CDATA #IMPLIED
+	savePrefix CDATA #IMPLIED
+	saveResponse CDATA #IMPLIED
+	soapAction CDATA #IMPLIED
+	url CDATA #REQUIRED
+	username CDATA #IMPLIED
 >
 <!ELEMENT pdfToBookmarksFilter EMPTY >
 <!ATTLIST pdfToBookmarksFilter
